Add: messages in dashboard. added templates, urls, and views for dispalying messages in dashboard.html fix: dashboard and others: fixed styles
