+++ account/views.py
@@ -1,6 +1,8 @@
 from django.contrib.auth.decorators import login_required
+import json
+from django.http import HttpResponse
 from django.shortcuts import render
-from properties.models import Property
+from properties.models import Property, Contact
 
 
 @login_required
@@ -21,3 +23,17 @@ def favorite_card(request):
     following = user.rel_from_set.all()
     properties = [prop.property_followed for prop in following]
     return render(request, 'properties/property_card_scroll.html', {'properties': properties})
+
+def messages(request):
+    user = request.user
+    props = user.my_properties.all()  # get user's properties
+    messages = [prop.contacts.all() for prop in props if len(prop.contacts.all()) > 0]  # get contacts for all user's properties with messages
+    return render(request, 'properties/messages_list.html', {'messages': messages})
+
+@login_required
+def message_detail(request, mess_id):
+    message = Contact.objects.get(id=mess_id)
+    if request.method == 'POST':
+        message.delete()
+        return HttpResponse(json.dumps(1), content_type='application/json')
+    return render(request, 'properties/message_details.html', {'mess': message})
