+++ properties/templates/properties/contact.html
@@ -332,7 +332,6 @@
           }else if( cont == 3 ){
             $(".liBtnCuenta").removeClass('animated heartBeat');
             clearInterval( id );
-            console.log(cont)
           }
         },500);
     });
